refactor: 필요없는 주석 제거 및 패키지 이동
