+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -2,7 +2,11 @@ package kitchenpos.menu.domain;
 
 import kitchenpos.menu.exception.MenuProductException;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
