+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,6 +1,10 @@
 package kitchenpos.product.domain;
 
-import javax.persistence.*;
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.math.BigDecimal;
 import java.util.Objects;
